chore: Update verifyUser middleware to use getCookies
This commit updates the `verifyUser` middleware in the `frontend/src/middlewares/verifyUser.ts` file. It replaces the usage of the `getCookie` function with the `getCookies` function from the `cookies-next` library. It also adds a console log to display the cookies. This change was made to improve the functionality of the middleware.